requirements.txt
alabaster==0.7.12
awscli==1.22.41
Babel==2.9.1
botocore==1.23.41
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.3
coverage==6.2
docutils==0.15.2
flake8==4.0.1
idna==3.3
imagesize==1.3.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.22.0
packaging==21.3
pandas==1.4.0
pip==21.3.1
pyasn1==0.4.8
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
rsa==4.7.2
s3transfer==0.5.0
scikit-learn==1.0.2
scipy==1.7.3
setuptools==60.5.0
six==1.16.0
sklearn==0.0
snowballstemmer==2.2.0
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
src==0.1.0
threadpoolctl==3.0.0
urllib3==1.26.8
wheel==0.37.1
